src/runner_test.go
package main

import (
	"os"
	"os/exec"
	"testing"
)

func TestProcessSignedScript(t *testing.T) {
	shouldVerifyYaml := false
	shouldDoInsightsCoreGPGCheck := false
	temporaryWorkerDirectory := "test-dir"
	config = &Config{
		VerifyYAML:               &shouldVerifyYaml,
		TemporaryWorkerDirectory: &temporaryWorkerDirectory,
		InsightsCoreGPGCheck:     &shouldDoInsightsCoreGPGCheck,
	}

	defer os.RemoveAll(temporaryWorkerDirectory)

	// Test case 1: verification disabled, no yaml data supplied = empty output
	yamlData := []byte{}
	expectedResult := ""
	result := processSignedScript(yamlData)
	if result != expectedResult {
		t.Errorf("Expected %q, but got %q", expectedResult, result)
	}

	// Test case 2: verification disabled, yaml data supplied = non-empty output
	yamlData = []byte(`
vars:
    _insights_signature: "invalid-signature"
    _insights_signature_exclude: "/vars/insights_signature,/vars/content_vars"
    content: |
        #!/bin/sh
        echo "$RHC_WORKER_FOO $RHC_WORKER_BAR!"
    content_vars:
        FOO: Hello
        BAR: World`)
	expectedResult = "Hello World!\n"
	result = processSignedScript(yamlData)
	if result != expectedResult {
		t.Errorf("Expected %q, but got %q", expectedResult, result)
	}

	// FIXME: This is false success because verification fails on missing insighs-client
	// Test case 3: verification enabled, invalid signature = error msg returned
	shouldVerifyYaml = true
	shouldDoInsightsCoreGPGCheck = true
	expectedResult = "Signature of yaml file is invalid"
	result = processSignedScript(yamlData)
	if result != expectedResult {
		t.Errorf("Expected %q, but got %q", expectedResult, result)
	}
}

func TestVerifyYamlFile(t *testing.T) {
	shouldVerifyYaml := false
	shouldDoInsightsCoreGPGCheck := false

	config = &Config{
		VerifyYAML:           &shouldVerifyYaml,
		InsightsCoreGPGCheck: &shouldDoInsightsCoreGPGCheck,
	}
	// Test case 1: verification disabled
	expectedResult := true
	result := verifyYamlFile([]byte{})
	if result != expectedResult {
		t.Errorf("Expected %v, but got %v", expectedResult, result)
	}

	// Test case 2: verification enabled and verification succeeds
	shouldVerifyYaml = true
	// FIXME: This should succedd but now verification fails on missing insighs-client
	// We also need valid signature
	expectedResult = false
	result = verifyYamlFile([]byte("valid-yaml"))
	if result != expectedResult {
		t.Errorf("Expected %v, but got %v", expectedResult, result)
	}

	// FIXME: Valid test case but fails because of missing insights-client
	// Test case 3: sverification is enabled and verification fails
	// shouldVerifyYaml = true
	expectedResult = false
	result = verifyYamlFile([]byte("invalid-yaml")) // Replace with your YAML data
	if result != expectedResult {
		t.Errorf("Expected %v, but got %v", expectedResult, result)
	}
}

// Function to check if one string slice is a subset of another
// Simple compare isn't enough because environment can change during execution
func areStringSlicesSubset(subset, full []string) bool {
	for _, s := range subset {
		found := false
		for _, f := range full {
			if s == f {
				found = true
				break
			}
		}
		if !found {
			return false
		}
	}
	return true
}

func TestSetEnvVariablesForCommand(t *testing.T) {
	testCases := []struct {
		name                string
		variables           map[string]string
		expected            []string
		anotherCmdVariables map[string]string // Same variables with different values for another command
	}{
		{
			name: "SettingVariables",
			variables: map[string]string{
				"VAR1": "value1",
				"VAR2": "value2",
			},
			expected: []string{
				"RHC_WORKER_VAR1=value1",
				"RHC_WORKER_VAR2=value2",
			},
			anotherCmdVariables: map[string]string{
				"VAR1": "another_value1",
			},
		},
		{
			name:      "EmptyVariables",
			variables: nil,
			expected:  nil, // Expect no changes to command's environment in this case
			anotherCmdVariables: map[string]string{
				"VAR2": "another_value2",
			},
		},
		// Add more test cases as needed
	}

	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			originalEnv := os.Environ()
			// Create the first dummy command
			cmd := exec.Command("echo", "Hello, World!")
			anotherCmd := exec.Command("echo", "Bye, World!")

			// Call the functions to set env variables for the commands
			setEnvVariablesForCommand(cmd, tc.variables)
			setEnvVariablesForCommand(anotherCmd, tc.anotherCmdVariables)

			// Check if the global environment variables are unchanged
			if !areStringSlicesSubset(originalEnv, os.Environ()) {
				t.Error("Global environment variables have been modified.")
			}

			// Check if the first command's environment variables have been set correctly
			if !areStringSlicesSubset(cmd.Env, append(os.Environ(), tc.expected...)) {
				t.Errorf("Command's environment variables are incorrect. Got: %v, Expected: %v", cmd.Env, append(os.Environ(), tc.expected...))
			}

			// Check if the second command's environment variables are NOT same as for first command
			if areStringSlicesSubset(anotherCmd.Env, append(os.Environ(), tc.expected...)) {
				t.Errorf("Command's environment variables are incorrect. Got: %v, Expected: %v", cmd.Env, append(os.Environ(), tc.expected...))
			}
		})
	}
}
